+++ database/migrations/2019_05_02_141239_create_playlists_table.php
@@ -16,8 +16,9 @@ class CreatePlaylistsTable extends Migration
     public function up()
     {
         Schema::create(
-            'playlists', function (Blueprint $table) {
-                $table->increments('id');
+            'playlists',
+            function (Blueprint $table) {
+                $table->bigIncrements('id');
 
                 $table->timestamps();
             }
